A fix related to https://github.com/sparanoid/live-dl/issues/7 after I cherry-picked.
When checking private videos/streams (for example, member only streams), cookies are required here.
But the original code does not work with cookie files. Therefore, I modified the perl regex to make it work.
My docker image can ensure that there must be a cookie file passed in, so there is no need to check in my case.
**(Please note that if no cookie file is passed in, it will not work properly. And you should use the original code.)**